*: fix some dumb printf format warnings

+++ sharpd/sharp_zebra.c
@@ -212,8 +212,8 @@ static int route_notify_owner(ZAPI_CALLBACK_ARGS)
 		if (sg.r.total_routes == sg.r.installed_routes) {
 			monotime(&sg.r.t_end);
 			timersub(&sg.r.t_end, &sg.r.t_start, &r);
-			zlog_debug("Installed All Items %ld.%ld", r.tv_sec,
-				   (long int)r.tv_usec);
+			zlog_debug("Installed All Items %jd.%ld",
+				   (intmax_t)r.tv_sec, (long)r.tv_usec);
 			handle_repeated(true);
 		}
 		break;
@@ -228,8 +228,8 @@ static int route_notify_owner(ZAPI_CALLBACK_ARGS)
 		if (sg.r.total_routes == sg.r.removed_routes) {
 			monotime(&sg.r.t_end);
 			timersub(&sg.r.t_end, &sg.r.t_start, &r);
-			zlog_debug("Removed all Items %ld.%ld", r.tv_sec,
-				   (long int)r.tv_usec);
+			zlog_debug("Removed all Items %jd.%ld",
+				   (intmax_t)r.tv_sec, (long)r.tv_usec);
 			handle_repeated(false);
 		}
 		break;
